# Optimize serialization for nested structures