ref: #93 

```ruby
require 'benchmark/ips'
require 'aerospike'

data = {"i"=>{"76908"=>{19=>[1578568165, 20, 1578568165, 0], 4=>[1578568165, 20, 1578568165, 0], 40=>[1578568165, 20, 1578568165, 0], 9=>[1578568165, 20, 1578568165, 0], 49=>[1578568165, 20, 1578568165, 0]}}, "r"=>{"pwyx"=>{40=>[1578568165, 20, 1578568165, 0], 48=>[1578568165, 20, 1578568165, 0], 27=>[1578568165, 20, 1578568165, 0], 38=>[1578568165, 20, 1578568165, 0], 18=>[1578568165, 20, 1578568165, 0]}, "uwbt"=>{35=>[1578568165, 20, 1578568165, 0], 42=>[1578568165, 20, 1578568165, 0], 20=>[1578568165, 20, 1578568165, 0], 23=>[1578568165, 20, 1578568165, 0], 30=>[1578568165, 20, 1578568165, 0], 4=>[1578568165, 20, 1578568165, 0]}, "adur"=>{10=>[1578568165, 20, 1578568165, 0], 38=>[1578568165, 20, 1578568165, 0], 45=>[1578568165, 20, 1578568165, 0], 16=>[1578568165, 20, 1578568165, 0], 33=>[1578568165, 20, 1578568165, 0], 1=>[1578568165, 20, 1578568165, 0], 2=>[1578568165, 20, 1578568165, 0], 28=>[1578568165, 20, 1578568165, 0]}, "uzxb"=>{18=>[1578568165, 20, 1578568165, 0], 19=>[1578568165, 20, 1578568165, 0], 8=>[1578568165, 20, 1578568165, 0], 3=>[1578568165, 20, 1578568165, 0]}, "tazd"=>{39=>[1578568165, 20, 1578568165, 0], 26=>[1578568165, 20, 1578568165, 0]}}, "b"=>{}, "d"=>{}, "l"=>{"qy"=>{9=>[1578568165, 20, 1578568165, 0], 11=>[1578568165, 20, 1578568165, 0], 32=>[1578568165, 20, 1578568165, 0], 13=>[1578568165, 20, 1578568165, 0], 27=>[1578568165, 20, 1578568165, 0], 23=>[1578568165, 20, 1578568165, 0]}, "ta"=>{48=>[1578568165, 20, 1578568165, 0], 22=>[1578568165, 20, 1578568165, 0], 1=>[1578568165, 20, 1578568165, 0], 17=>[1578568165, 20, 1578568165, 0], 33=>[1578568165, 20, 1578568165, 0]}}, "o"=>{"rthcbjqd"=>{19=>[1578568165, 20, 1578568165, 0], 4=>[1578568165, 20, 1578568165, 0], 40=>[1578568165, 20, 1578568165, 0], 27=>[1578568165, 20, 1578568165, 0], 9=>[1578568165, 20, 1578568165, 0], 13=>[1578568165, 20, 1578568165, 0], 3=>[1578568165, 20, 1578568165, 0], 15=>[1578568165, 20, 1578568165, 0]}, "ojrtvlnm"=>{2=>[1578568165, 20, 1578568165, 0], 42=>[1578568165, 20, 1578568165, 0], 19=>[1578568165, 20, 1578568165, 0], 20=>[1578568165, 20, 1578568165, 0], 30=>[1578568165, 20, 1578568165, 0], 8=>[1578568165, 20, 1578568165, 0], 14=>[1578568165, 20, 1578568165, 0]}}, "v"=>{"hdzn"=>{39=>[1578568165, 20, 1578568165, 0], 30=>[1578568165, 20, 1578568165, 0], 10=>[1578568165, 20, 1578568165, 0], 1=>[1578568165, 20, 1578568165, 0], 33=>[1578568165, 20, 1578568165, 0], 24=>[1578568165, 20, 1578568165, 0], 19=>[1578568165, 20, 1578568165, 0], 6=>[1578568165, 20, 1578568165, 0]}}}

GC.disable

Benchmark.ips do |x|
  x.report('list value with hash') { Aerospike::ListValue.new([{ 's' => data }]).to_bytes }
  x.report('list value with one less hash') { Aerospike::ListValue.new([data]).to_bytes }
  x.report('just map value') { Aerospike::MapValue.new(data).to_bytes }
end
```

Before:

```
Warming up --------------------------------------
list value with hash    14.000  i/100ms
list value with one less hash
                        30.000  i/100ms
      just map value    64.000  i/100ms
Calculating -------------------------------------
list value with hash    157.353  (± 7.0%) i/s -    784.000  in   5.010336s
list value with one less hash
                        369.986  (± 5.4%) i/s -      1.860k in   5.044823s
      just map value    792.737  (± 7.4%) i/s -      3.968k in   5.041397s
```

After:

```
Warming up --------------------------------------
list value with hash   522.000  i/100ms
list value with one less hash
                       594.000  i/100ms
      just map value   590.000  i/100ms
Calculating -------------------------------------
list value with hash      5.421k (±12.7%) i/s -     26.622k in   5.000921s
list value with one less hash
                          5.577k (±10.0%) i/s -     27.918k in   5.062942s
      just map value      6.057k (± 5.9%) i/s -     30.680k in   5.083538s
```

35x speedup.